refactor dashboard with auth middleware
